update to 1.20 and no longer depend on glib

+++ src/main/java/me/gravityio/viewboboptions/VanillaOptions.java
@@ -30,8 +30,8 @@ public class VanillaOptions {
 
 
     public static void init() {
-        ALL_BOBBING_STRENGTH.setValue(ModConfig.INSTANCE.all_bobbing_strength());
-        HAND_BOBBING_STRENGTH.setValue(ModConfig.INSTANCE.hand_bobbing_strength());
-        CAMERA_BOBBING_STRENGTH.setValue(ModConfig.INSTANCE.camera_bobbing_strength());
+        ALL_BOBBING_STRENGTH.setValue((int) ModConfig.INSTANCE.all_bobbing_strength);
+        HAND_BOBBING_STRENGTH.setValue((int) ModConfig.INSTANCE.hand_bobbing_strength);
+        CAMERA_BOBBING_STRENGTH.setValue((int) ModConfig.INSTANCE.camera_bobbing_strength);
     }
 }
